apps/backend/requirements.txt
# Core runtime
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
PyYAML>=6.0.2
tenacity>=9.0.0
structlog>=24.4.0

# Google ADK + Gemini + Google Cloud
google-adk>=1.0.0
google-genai>=0.8.0
google-cloud-aiplatform>=1.71.0
google-cloud-logging>=3.11.0
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.18.0
google-cloud-firestore>=2.18.0
google-cloud-pubsub>=2.23.0
google-cloud-scheduler>=2.14.0
google-cloud-texttospeech>=2.17.0
google-cloud-speech>=2.27.0

# Firebase Admin
firebase-admin>=6.5.0

# LangChain / LangGraph
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-vertexai>=2.0.0

# Browser automation
playwright>=1.45.0

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48b0

# Crypto (Ed25519 device PoP)
PyNaCl>=1.5.0

# Testing / quality
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
ruff>=0.6.0
mypy>=1.11.0
